+++ packages/builders/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @nyxojs/builders
 
+## 0.2.5
+
+### Patch Changes
+
+- @nyxojs/rest@0.2.5
+
 ## 0.2.4
 
 ### Patch Changes